Adding new firebaseModels for firebase database + updating the context classes to connect to firebase Database

+++ HotelManager/ServiceLayer/Managers/ClientManager.cs
@@ -1,5 +1,6 @@
 ﻿
 using DataLayer;
+using System.Linq;
 
 namespace ServiceLayer
 {
@@ -12,16 +13,29 @@ namespace ServiceLayer
         }
         public ClientManager(ClientContext clientContext)
         {
-            this.clientContext = clientContext;
+            this.clientContext = clientContext ?? throw new ArgumentNullException(nameof(clientContext));
         }
         public async Task CreateAsync(Client client)
         {
+            if (client == null)
+                throw new ArgumentNullException(nameof(client), "Client cannot be null.");
+
+            await EnsureUniqueClientAsync(client);
+
             await clientContext.CreateAsync(client);
         }
 
         public async Task<Client> ReadAsync(Guid clientId, bool useNavigationalProperties = false, bool isReadOnly = true)
         {
-            return await clientContext.ReadAsync(clientId, useNavigationalProperties, isReadOnly);
+            if (clientId == Guid.Empty)
+                throw new ArgumentException("Invalid Client ID.", nameof(clientId));
+
+            var client = await clientContext.ReadAsync(clientId, useNavigationalProperties, isReadOnly);
+
+            if (client == null)
+                throw new KeyNotFoundException($"Client with ID {clientId} does not exist.");
+
+            return client;
         }
 
         public async Task<ICollection<Client>> ReadAllAsync(bool useNavigationalProperties = false, bool isReadOnly = true)
@@ -31,12 +45,35 @@ namespace ServiceLayer
 
         public async Task UpdateAsync(Client client, bool useNavigationalProperties = false)
         {
+            if (client == null)
+                throw new ArgumentNullException(nameof(client), "Client cannot be null.");
+
+            var existingClient = await clientContext.ReadAsync(client.Id, useNavigationalProperties, false);
+            if (existingClient == null)
+                throw new KeyNotFoundException($"Client with ID {client.Id} does not exist.");
+
+            await EnsureUniqueClientAsync(client, existingClient.Id);
+
             await clientContext.UpdateAsync(client, useNavigationalProperties);
         }
 
         public async Task DeleteAsync(Guid clientId)
         {
+            if (clientId == Guid.Empty)
+                throw new ArgumentException("Invalid Client ID.", nameof(clientId));
+
+            var existingClient = await clientContext.ReadAsync(clientId, false, false);
+            if (existingClient == null)
+                throw new KeyNotFoundException($"Client with ID {clientId} does not exist.");
+
             await clientContext.DeleteAsync(clientId);
+        }
+        private async Task EnsureUniqueClientAsync(Client client, Guid? existingClientId = null)
+        {
+            var existingClients = await clientContext.ReadAllAsync();
+
+            if (existingClients.Any(c => c.Email == client.Email && c.Id != existingClientId))
+                throw new InvalidOperationException($"A client with email {client.Email} already exists.");
         }
     }
 }
